added error message and doc in matlab wrapper to enforce N should be a power of 2
Now the doc of the matlab wrapper also notes the power of two restriction for all cases that use FFT. However it is checked in qdyn.m only when MESHDIM=1. If MESHDIM=2, the use of FFT is not set in qdyn.m but in constants.f90, thus the check should be done in the fortran code.