Re-structured the packing/unpacking of variables (#16)
Prior to the solver step, variables (theta, tau/v, etc.) are merged (packed) into a single variable and passed onto the solver subroutine. After the solver step, the resulting variable is split-up (unpacked) back into the original variables. This process is repeated on various locations in the code with some code duplication. The packing/unpacking procedure is now centralised as a subroutine in utils.f90