Implementation of thermal pressurisation (#28)
An improved implementation for high-velocity thermal pressurisation, based on Noda & Lapusta (2010). The diffusion of fluid and heat normal to the fault plane is solved in a discretised spectral domain, which is numerically stable for large time steps (as opposed to discretisation methods in the spatial domain, such as FEM, FD, etc.). This procedure allows for efficient modelling of frictional heating and associated pressurisation of the pore fluid.

This module is compatible with both the rate-and-state friction and the Chen-Niemeijer-Spiers (CNS) frameworks. When the CNS model is used, dilatancy hardening is accounted for through a dilatancy efficiency coefficient (set to 0 to ignore dilatancy hardening).

Fixed a small bug in `pyqdyn.py` in which multiple columns in the time series/snapshots output were named identically.